backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
pymongo
bcrypt
python-jose[cryptography]
passlib[bcrypt]
